requirements.txt
accelerate==1.4.0
came_pytorch==0.1.3
einops==0.8.1
diffusers@git+https://github.com/huggingface/diffusers
clip@git+https://github.com/openai/CLIP.git
gradio==5.20.0
ipdb==0.13.13
h5py==3.13.0
matplotlib
mmcv==1.7.2
numpy==2.1.2
scipy==1.15.2
scikit-learn==1.6.1
scikit-image==0.25.2
omegaconf
opencv-python==4.11.0.86
patch_conv==0.0.1b0
pycocotools==2.0.8
pillow
pandas
pyrallis
pytorch-fid==0.3.0
sentencepiece==0.2.0
tensorboard==2.19.0
termcolor
timm==0.6.13
torch==2.4.0
torchvision==0.19.0
tqdm
transformers==4.49.0
lpips==0.1.4
